Integrate "on DS laptop" changes from BattleCry
The goal here was to preserve the changes that were made in the heat of
battle. The problem was that the DS laptop seems to have started the
battle with an out-of-sync version of the code, so when we tried to
integrate the changes, conflict errors were thrown. I *think* I have
captured all the pertinent changes in these code modifications. These
need to be TESTED THOROUGHLY!
What I *know* is that we have to move off of Git and use something
simpler as our VCS - we cannot handle (nor do we really need) the
complexity of Git. I propose we move to Subversion this summer...